requirements.txt
dht11
RPi.GPIO